Decouple Tag and Input from Form, transfer coupling to Formatter
Tag and Input should be independent for ease of use.  The Formatter
needs access to the Form so that the Form can affect the conversion
from Input to Tag.  This isn't currently used, but it may be later
for example to support a :namespace option on Forms.

One affect of these changes is that Tag and Input objects can no
longer serialize themselves, but I don't expect that to be a
problem.