Major structural changes
Make Input and Tag keep a reference to the Form.  Have Form#input and
Form#tag return Input and Tag instances, and have Input#to_s and
Tag#to_s return serialized versions of the tags, which they can now
do.  This means that when Forme is used as a standalone DSL, it
doesn't need to serialize everything twice.

No longer have the serializer and formatter procs take a form
argument, since it's now an attribute of the tag/input argument.

Add Form#_input, Form#_tag, and Tag#tag for creating Input/Tag
instances associated to the related form.

Currently, the only ugly part is having a transformer that takes
an array now needs to scan it for a tag instance in order to get
a form reference to use to create other tags.  That could go away,
but it requires that instead of an array, we add something like
Forme::TagArray that stores an array and a form reference, and
change the API so that plain array instances are automatically
wrapped in a TagArray before calling the transformer.