Handle implicitly emitted code in the Sinatra integration
Previously, things like :legend in the default InputsWrapper
and hidden primary key fields in nested attributes didn't
work correctly in the Sinatra integration, because of the way
it currently works by emitting code into the output as it runs
(as opposed to taking control of the whole output from sinatra
and adding giving sinatra back control when it is finished).

Basically, for tags that use blocks but contain implicitly
created tags/inputs, the implicitly created tags/inputs
need to be passed to emit so they can be injected into the
template output.

Also, don't yield self from inputs, as it doesn't gain anything.