Refactor transformer handling
Instead of separate hash per transformer class, use a single
hash with subhashes per type.  Make register_transformer a
Forme module method and take the transformer type as first
argument.

Add Form#transform and #transformer to make it easier to
handle transforms.  Remove the TransformerMap module as it
is no longer needed.  Remove the wrapper class as it is
simple enough to be handled by regular procs.