Make Tag instances use nil children by default
Before, it would create a TagArray for tags without children,
which is pretty wasteful.  This makes it so that tags without
children use nil.

Note that this requires API changes since you cannot assume that
you can just iterate over a tag's children (you have to check
for the children's existance first).