Make error_handler, labeler, and wrapper take tag and input
This unifies handling of the error_handler, labeler, and wrapper
transformers so they call take the same types of arguments. This
removes the need to scan TagArrays looking for tags which was
previously done by some of the transformers.  It will enable the
ability for those transformers to modify the transformations
based on the initial input, which previously was not possible.

Remove TagArray#tag and add Input#tag, and have all of the
transformers call #tag on the input argument instead of the
tag argument.

Add Input#merge_opts for merging new options into the input's
opts without modifying the input's opts hash.  This is necessary
for cases where the formatter needs to modify the Input's opts
in order for correct behavior of another transformer, currently
used by the date to multiple select options code.

Only turn arrays of tags into TagArrays at the very end of the
formatter, since the other transformers should no longer care.
TagArray is still needed, but hopefully it can go away in the
future.