Replace TagArray class with Serialized module
Instead of wrapping all arrays in TagArray instances, just
extend the arrays with a module that overrides to_s to use
the form's serializer.

This is currently only needed if Form#input can return an array of
inputs, which the Sequel plugin does for the associations as
multiple radios/checkboxes option.