Don't copy all input options to tag attributes, use an :attr hash
This cleans up the code and makes it so that we don't need to
do so much hash mangling.

Still copy the common options such as name, id, placeholder, class,
and value, to make the API nice.  Combine class entries instead of
overwriting.

Support a :data option to create data-* attributes easily.