Make the demo site read-only if DATABASE_URL is given
This way if the repository is cloned, the demo site can
be run locally for testing of updates.  However, allowing
updating on a public demo site is problematic because multiple
users can be making changes as once.  As the purpose of the site
is to show off Forme, allowing updating is not necessary.