Add Rails integration
This is cleaner in some ways than the Sinatra integration, as
Rails templates hack ERB to allow <%= to be used with blocks,
and Rails already has code that allows you to switch output
buffers.  It still is fairly clunky in regards to what gets
emitted versus what gets returned, but it does pass the same
specs as the sinatra support.