Make html_usa serializer convert date/datetime inputs to text for better compatibility
The HTML5 date/datetime inputs are specified to use ISO specific
values, and the html_usa serializer changes the values to be
formatted American-style middle-endian (month/day/year).  This
breaks compatibility with some HTML5 implementations, such as
Chrome, which shows the input as empty if the value is middle
endian.  The html_usa style was designed for browsers that treat
date/datetime inputs as text inputs, so automatically convert
date/datetime inputs to text inputs in the serializer.