Move namespace handling from Sequel plugin into the core
While namespace handling is mostly useful for model forms, non-model
forms could also benefit from it.

It's common to use the same base values for the name and id
attributes, so add a :key option for that.  Also support a :key_id
option so that the id can vary even if the name stays the same,
useful for multiple checkboxes or radio buttons.