Support radioset and checkboxset Input types
This abstracts the code used to display associations as radio
buttons (many_to_one) or checkboxes (one_to_many), and moves it
into the Forme core, where it can be used by anyone.

radioset and checkboxset share the :options option parsing with
the select type, which simplifies a lot of the code.