Demo site changes
Always show form_opts and form_attr on the example pages.

Show the last album instead of the first, since the first album
is self-titled.

Don't show tracks association in the basic forms, since it doesn't
make logical sense, and the tags association shows all of the
same features.

Mention on the front page that the forms aren't submittable, and
that you should view the source to see the markup differences.