Add Form.with_opts method to override opts for the given block
This takes quite a bit of refactoring because in the default case
(without templates), formatting is not done at the time of
Input creation.  So Inputs need to copy a reference to the Form
opts when they are created, and use that later when formatting.

Since the transformers for a form are no longer static, don't
set instance variables for them, just look them up in the opts.
The accessor methods for the transformers can go away as well.
The one exception is the serializer, which needs to remain
static.

Additionally, the Form#format and #serialize methods can be
removed as well, since nothing internally calls them, all
access goes through the transformer.