Store namespaces in Form opts
Instead of storing namespaces as a separate attribute, store them
in the opts, so that you can use with_opts to change the temporarily.

Additionally, add a with_obj method that takes a new object and
additional namespaces, and replaces the object and updates the
namespace temporarily.  This appends the given namespaces to the
current namespace.

Remove the Form#push_namespace and #pop_namespace private methods.