Make explicit labeler use label-before and label-after classes specifying the label position
Forme doesn't automatically add colons when using the explicit
labeler, it assumes they will be added via a CSS :after content
setting.  However, in general you'll only want to add colons if
the label comes before the input, not after.  So this makes it
easy to use the appropriate CSS selector to only add the colons
where they should be added.