Minor changes to previous commit
Remove unnecessary model.include(self), and use .send(:include)
instead of .include.

Simplify the module structure in the forme_i18n plugin.

Require i18n >=0.7.0 to run the specs, as the code does not
appear to work in 0.6.0.

Skip the i18n related specs if the library is not installed,
and just print a warning, similar to the other integrations.