Add forme_set Sequel plugin, for handling the intake of submitted params based on form fields
This plugin tracks which inputs have been added for an object, and
adds a forme_set model instance method that will take a parameter
hash and only set the values that match one of the inputs added via
the form.  Additionally, when filtering associated objects when
displaying the form (or when using filters on associations), adds
validations that the value submitted by the form matches one of the
options in the form.

This doesn't handle all possible cases.  For example, it's possible
to use non-model backed fields on a model form via :obj=>nil.
forme_set will just ignore those inputs, but you'll still be
able to set them manually.