Use thead/tbody tags for table inputs_wrapper transformer and Sequel plugin subform :grid option
Without thead tags, if you are using the :labels options, you can't
differentiate the header rows from the input rows.

While this change isn't really necessary if the :labels option is
not used, since the tbody tag can be safely omitted in that case,
it's simpler to just always use tbody.