Allow overriding input type when using an associated object
Overriding the input type when an associated object is used is currently
only possible when using Sequel model instances (and the Sequel plugin
loaded). When using a custom object, the input type is always "text".

This commit adds the :type option for overriding the input type (just
like the :type option that the Sequel plugin supports). The field value
from the object will be set for all input types except "file", as for
file fields it semantically doesn't have any meaning and generally it's
not desirable.