Add Roda forme_route_csrf plugin using route_csrf plugin for request-specific CSRF tokens
Also add forme_default_request_method as a method to check for
object specific forms.  If defined, calls the method and uses
the resulting value as the default method for the form (if no
method is specified).  This is now used in the Sequel plugin.

This also fixes checking of form tag attributes for whether to
add a CSRF token, working with both symbol and string keys now.