Use div with nested p tags instead of spans for readonly textarea inputs
For text inputs, since they are single lines, using a span is fine.

However, since textareas are designed to be used for multiline input,
using a span results in visual issues because it doesn't preserve
the line breaks.  Using a div with nested p tags for each paragraph
and br tags for each individual line break is probably the closest
way to get a similar visual style in readonly formatting.