Support :fieldset wrapper and :legend labeler, can be used for accessible radioset/checkboxset
Also, add support for :tag_label_attr option, to allow different
attributes for option labels compared to the set_label.  For
partial backwards compatibility, if :tag_label_attr is not given
and :label_attr is given, :label_attr will be used for both the
option labels and set label. This backwards compatibility is
only partially, because previously there was not a way to set
the attributes for the set label.  Backwards compatibility in
terms of using :label_attr for the set_label was broken when
custom :labelers were allowed for sets.