Add Roda forme_set plugin, using HMACed form metadata to automatically handle submitted form parameters
This builds on top of the Sequel forme_set plugin, for even easier
handling of form submissions.  To make sure the submission can't
be forged, it uses an HMAC of the submitted form data.  It makes
sure the metadata is only valid for the specific form by including
the CSRF token in the metadata.