Update bs3 formatter
The generic formatter's `#normalize_options` method is copied over bs3's
formatter with the only purpose to remove the `error` class that is
added to the input. The copied version had become staled and it lacked two
new features:

- Errors defined at the form level through the `:errors` key (see
  https://github.com/jeremyevans/forme/pull/32).
- The addition of accessibility attributes `aria-invalid` and
  `aria-describedby` (see
  https://github.com/jeremyevans/forme/commit/d070e6a1508c7427bc29581925f6a462e835b971).

This commit updates the copied version, updates tests to contain `aria-`
attributes and add tests for the `errors:` option.