Refactor internals, switch template integration to use delegation
Internally, forme was kind of a mess to support the template
integration for the multiple web frameworks Forme supports.
Methods like #emit, #tag_, and #raw_output were defined on
Forme::Form, even though there was no need for those outside of
template integration.  Integrating support for new templates
was a huge pain, handled mostly by getting the tests passing using
trial and error, adding hacks as needed so the new template
integration would work without breaking the existing template
integrations.  As the tests don't cover all possible usage, this
could result in breakage in applications, which happened during
the recent addition of the erubi capture template support.

The main change to the internals outside of the template integrations,
other than removing methods no longer needed, is that Forme::Form
now has an internal HTML buffer of the HTML added to the form.  The
internal abstract tree is now eagerly consumed instead of lazily
consumed to produce HTML, which is added to the form's buffer.  The
input, inputs, and tag methods now return all content added to the
form's buffer during the call to the method, handling cases where
fields are added internally.  Forms no longer store a nested tree
of objects.  Inside of using a nesting, where objects added to the
form are added to the current nested object, places where nesting
was used (via blocks) now adds a opening tag to the HTML buffer,
yields to the block (which will likely result in the HTML buffer
being updated while inside the tag), then adds a closing tag to the
HTML buffer.

This uses a delegation approach to template integrations instead of
having them subclass Forme::Form (well, the Roda forme_set plugin
integration still subclasses, but not for template integration).
The delegation then only needs to override a few methods
on the delegate object in order to handle the template integration.
In general, the template integrations are much simpler now, and share
most of the same code, using a new forme/template file.

This fixes the failing subform specs in the previous commit.

This is not completely backwards compatible, and will result in a major
version bump.  However, it's expected that most applications using
Forme will not need changes.  A few known incompatibilities:

* Most Forme::Form methods to generate tags return HTML strings
  instead of returning Forme objects that will be converted to
  HTML strings via to_s.

* When calling tag and passing a proc or method as the children
  (third argument), the proc or method should call methods that
  just return children, instead of methods that modify the form
  directly (e.g. calling tag on the argument passed to the
  proc method, instead of calling form.tag on the argument).

* Mutating the string returned by Forme::Form#to_s will mutate
  future the internal state of the form.  You'll need to avoid
  doing that.

* The following methods have been removed from Forme::Form:

  * nesting
  * tag_
  * raw_output
  * emit

* _forme_form_options private method used in the template
  integrations now takes the object, attributes, and options,
  instead of just the options.