Add :before and :after Form options, as an alternative way to implement hooks
This also adds :_before and :_after internal options, which are used to
implement the template integrations.

This offers an alternative for the current hidden tags support, which
will be deprecated and later removed.

Simplify the template integrations to use the _before and _after hooks.
Add Form#form_tag_attributes so the _before and _after hooks can get
access to the attributes.