Get to 100% branch coverage
As is typical, getting to 100% branch coverage uncovered multiple bugs.

The largest bug is that in the bs3 support, Forme was creating
aria-describedby attributes referencing ids that didn't exist,
because the bs3 error handler didn't handle them.

The was a minor bug in that aria-describedby was not added
automatically if an :error_id option was given manually.

I also found that boolean inputs in the Sequel forme plugin
didn't correctly handle explicit nil/false :value options in
all cases.