Minor tweaks to the Bootstrap 5 support
Forme.attr_classes_after has been added, which uses the current
classes in attr after the given classes when setting the class
attribute.  This was because there was a lot of test breakage
if just using Forme.attr_classes. Unless the order of classes
matters, it would be best to change the bs5 support to use
Forme.attr_classes and fix the breakage.