Support emit: false option for non-rails template forms allowing block based form use without appending to template
This is useful when calling methods that need to call form with
blocks, but where the block is not yielding control back to the
template.  This will allow support for CSRF tokens and forme_set
metadata for such forms, which did not work correctly before as
as mostly empty form would be appened to the template.