Omit closing / for void element tags
When Forme was originally created, xhtml compliance was considered
important, as html5 had not yet been released. Considering that
html5 is now over 10 years old, and the vast majority of
applications using Forme are likely using html5 and not xhtml,
it's seems better to drop support.

If someone really wants xhtml support, I'm open to adding it back
via an option.